+++ cookbooks/thomson_reuters_endnote/recipes/mac.rb
@@ -19,15 +19,36 @@ dmg_package "Thomson Reuters EndNote X7" do
   version "17.0.1"
 end
 
-# Install custom pkg that includes Word plugin
-dmg_package "Thomson Reuters EndNote X7 Extras" do
-  app "EndNote X7 Extras"
-  volumes_dir "EndNote X7 Extras"
-  dmg_name "endnote_x7_extras"
-  source "http://ims-chef.wesleyan.edu/os_x/thomson_reuters_endnote/endnote_x7_extras.dmg"
-  checksum "4a7351d6c7b6d1b0afde6101a2b3fcb0d660083a4d785d1905470e1bde88fdf7"
-  action :install
-  type "pkg"
-  package_id "com.endnote.x7.extras"
-  version "17.0.1"
+# Install the license file inside the EndNote directory
+cookbook_file "/Applications/EndNote X7/.license.dat" do
+  mode 0644
 end
+
+# Copy the plug-in from a directory inside EndNote to the Word 2011 plug-in directory.
+execute "EndNote X7 Word Plug-in" do
+  command "rsync --recursive --links --perms --executability --owner --group --times '/Applications/EndNote X7/Cite While You Write/EndNote CWYW Word 2011.bundle' '/Applications/Microsoft Office 2011/Office/Startup/Word'"
+end
+
+# Create directory for the EndNote Pages plug-in
+directory "/System/Library/Application Support/ResearchSoft/EndNote/Plugins" do
+  mode 00755
+  action :create
+  recursive true
+end
+
+# Copy the EndNote Pages plug-in.
+execute "EndNote X7 Pages Plug-in" do
+  command "rsync --recursive --links --perms --executability --owner --group --times '/Applications/EndNote X7/Cite While You Write/PagesEndNote.bundle' '/System/Library/Application Support/ResearchSoft/EndNote/Plugins/'"
+end
+
+# Copy the EndNote service to the default user profile.
+execute "EndNote X7 Service" do
+  command "rsync --recursive --links --perms --executability --owner --group --times '/Applications/EndNote X7/Services/ENService.app' '/System/Library/User Template/English.lproj/Library/Services/'"
+end
+
+# And finally, copy the EndNote plist file to the default user profile to suppress first launch dialogues.
+cookbook_file "/System/Library/User Template/English.lproj/Library/Preferences/com.ThomsonResearchSoft.EndNote.plist" do
+  mode 0600
+end
+
+