+++ cookbooks/thomson_reuters_endnote/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file is used to list changes made in each version of thomson_reuters_endnote.
 
+## 0.2.0:
+
+* Added code to copy Word plug-in to proper directory.
+* Added code to copy Pages plug-in to proper directory.
+* Added code to seed plist file and Service app into default user profile to surpress first-run dialogues.
+* Added code to seed license.dat file for registration.
+
 ## 0.1.0:
 
 * Initial release of thomson_reuters_endnote